Add LOC(Lines of Code) per language (#18)
Add LOC language for analysis

Change scripts config.json to be called projects.json. It gives a more clear idea of what the config is for.
Add segment attribute to analyses table. This will allow a split on different themes within an analysis like types of languages.

The loc languages processor will change git repos, this means that it needs to be cleaned up on failure(changed back to the main branch). 
In order to do this a cleanup method has been added and it is called on Critical Exceptions as well as Sighup and sigint.